refactoring, moving fee history related code to same file
