+++ crates/transaction-pool/src/lib.rs
@@ -101,7 +101,6 @@ use reth_provider::StateProviderFactory;
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::mpsc::Receiver;
 
-pub mod cache;
 mod config;
 pub mod error;
 mod identifier;
